Add Value Error when value is less than 1
Add the Vlaue error less than 1 because, numbers less than or equal to 0 is not accepted. I do agree that it might cause an issue in logic and <=0 should have been used. Note: Dangling if else issue- it will not cause any problem in this case as the rest of the program is only executed when there is no value Error. raising a value error halts the entire program.